fix: bundle script with dependencies
