+++ action.yml
@@ -14,4 +14,4 @@ inputs:
     required: false
 runs:
   using: 'node24'
-  main: 'action.js'
+  main: 'dist/index.js'
