+++ package.json
@@ -1,6 +1,15 @@
 {
     "dependencies": {
-        "github-label-sync": "3.0.0",
-        "@actions/core": "^1.10.0"
-    }
+        "@actions/core": "^1.10.0",
+        "github-label-sync": "3.0.0"
+    },
+    "devDependencies": {
+        "@rollup/plugin-commonjs": "^28.0.6",
+        "@rollup/plugin-node-resolve": "^16.0.1",
+        "rollup": "^4.46.2"
+    },
+    "scripts": {
+        "build": "rollup --config ./rollup.config.js"
+    },
+    "type": "module"
 }
